requirements.txt
# ----------------------
# Django
# ----------------------
Django==1.3.1
-e git+https://github.com/kmike/django-qsstats-magic.git#egg=django-qsstats-magic


# ----------------------
# Database
# ----------------------
psycopg2==2.4.1


# ----------------------
# Testing
# ----------------------
mock
coverage
django-test-coverage
mocker


# ----------------------
# Utilities
# ----------------------
python-dateutil==1.5
